Preparing to refactor TrapidInterproParser
